+++ internal/cmd/cmd.go
@@ -23,6 +23,7 @@ var (
 	flagConfig  string
 	flagPort    string
 	flagCors    bool
+	flagRegex   bool
 	flagDelay   int64
 	flagBaseApi string
 )
@@ -55,6 +56,8 @@ func Execute() {
 	serveCmd.PersistentFlags().StringArray("response-exec", []string{}, "endpoint response exec")
 	serveCmd.PersistentFlags().StringArray("middleware", []string{}, "middleware")
 	serveCmd.PersistentFlags().StringVarP(&flagBaseApi, "base", "b", "", "base API")
+
+	wipeHeadersCmd.PersistentFlags().BoolVar(&flagRegex, "regex", false, "enable regular expression for seaching")
 
 	if err := rootCmd.Execute(); err != nil {
 		fmt.Fprintln(os.Stderr, err)